Supervisor Functionalities and Fixes
DATABASE CHANGES:
+ NEW SupervisorRemarks for Day_Record
- CHANGED Datatype for OverallGrading in Month_Record to decimal

SUPERVISOR:
+ Grade Student Journal
+ View Student Journal
+ View Student Information

FIXES:
+ Changed getGrade javascript function in grading and review pages for LO and SUPERVISOR to fix error
+ Changed rendering of checked competencies to ensure unchecked competencies are unchecked.

--PLEASE REMIGRATE--

+++ E-Internship Journal/Models/Month_Record.cs
@@ -23,7 +23,7 @@ namespace E_Internship_Journal.Models
         public int? TechnicalGrading { get; set; }
         public int? IndependenceGrading { get; set; }
         public int? PerformanceGrading { get; set; }
-        public int? OverallGrading { get; set; }
+        public decimal? OverallGrading { get; set; }
         public string OverallFeedback { get; set; }
         
         public Internship_Record InternshipRecord { get; set; }
